% tinyrox says don't edit this manually, but it can't stop you!
\name{is_uri}
\alias{is_uri}
\title{Check if a string matches the pattern of a Spotify URI}
\usage{
is_uri(s)
}
\arguments{
\item{s}{String to check}
}
\value{
Returns a boolean.
}
\description{
Check if a string matches the pattern of a Spotify URI
}
