% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{pitch_class_lookup}
\alias{pitch_class_lookup}
\title{Pitch class notation lookup}
\format{
An object of class \code{character} of length 12.
}
\usage{
pitch_class_lookup
}
\description{
Pitch class notation lookup
}
\keyword{datasets}
