% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{scopes}
\alias{scopes}
\title{Valid scopes}
\format{
An object of class \code{character} of length 19.
}
\usage{
scopes
}
\description{
Vector of valid scopes for spotifyr::get_authorization_code()
}
\keyword{datasets}
