% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{tinyspotifyr-package}
\alias{tinyspotifyr}
\alias{tinyspotifyr-package}
\title{\packageTitle{tinyspotifyr}}
\description{\packageDescription{tinyspotifyr}}
\section{Package Content}{\packageIndices{tinyspotifyr}}
\author{\packageAuthor{tinyspotifyr}}
\section{Maintainer}{\packageMaintainer{tinyspotifyr}}
\keyword{internal}
