% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyspotifyr.R
\docType{package}
\name{tinyspotifyr}
\alias{tinyspotifyr}
\title{\code{tinyspotifyr} package}
\description{
A Quick and Easy Wrapper for Spotify's Web API
}
\details{
See the README on
\href{https://github.com/troyhernandez/tinyspotifyr#readme}{GitHub}
}
