% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_uri}
\alias{is_uri}
\title{Check if a string matches the pattern of a Spotify URI}
\usage{
is_uri(s)
}
\arguments{
\item{s}{String to check}
}
\value{
Returns a boolean.
}
\description{
Check if a string matches the pattern of a Spotify URI
}
